[ios] Remove kDestroy type and add kSelectionOnly type to WebStateListChange
This CL removes kDestroy from WebStateListChange and adds kSelectionOnly to WebStateListChange. This CL also adds a comment/a TODO comment to the kSelectionOnly arm.
We decided not to replace WebStateListDestroyed() with WebStateListChanged() because WebStateListDestroyed() is not compatible with other operations like an insertion. WebStateListDestroyed() is called when the entire WebStateList is destroyed, although other operations are called when one WebState is updated.
Instead, we add kSelectionOnly in this CL to replace WebStateActivatedAt() with WebStateListChanged() in another CL.

See "WebStateListChange class" section for more details (internal document):
https://docs.google.com/document/d/1I1QUZAexM1eMI2l2oriZl5GARrtofNI0SxHUPB4h-YE/edit?usp=sharing

Bug: 1442546
Change-Id: Ie1629a91efeaa8ac9220181a70536dd260d4031b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4607449
Reviewed-by: Federica Germinario <fedegermi@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158195}