Make SelectFileDialogParams a member of FileSystemChooserTest
Avoid a spurious dangling ptr warning about out-of-scope locals
by moving this from the stack into the test class itself where
it will persist until TearDown() clears the unowned reference.

In turn, this created a lifetime issue between the reference to
the GURL in SelectFileDialogParams and the object which owns it.
Storing this by value decouples the params, and copying a GURL
is relatively cheap since it doesn't reparse.

-- Add a comment about lifetime to ui::SelectFileDialog::SetFactory()
   while at it.

Bug: 1401495
Change-Id: I3361347a1b3f78c94c053254131a47feb1be5610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4616045
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158227}