[DIPS] Treat Successful Web Authn assertions as user interaction
This CL pipes into DIPS successful web Authn assertions signals to be
considered as user interaction by DIPS for the top level frames.

This is a first to a set of two, the following CL will pipe this same
signal for the Android Platform.

Bug: 1446678
Change-Id: I272d7a97fce5db1a1eea2b00cfeeb7f3e6843fa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4590609
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Jonathan Njeunje <njeunje@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1158235}