[//media][Mac VDA] Correctly check for null `config_.hdr_metadata`
[1] inadvertently flipped a check of `config_.hdr_metadata` being empty
to it being *non-empty*, resulting in it not being populated before
being accessed. This CL fixes the issue.

To test: Download the video in the linked bug and drag it into the
browser. The video should play and there should be no GPU process crash.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4596056

Bug: 1454723
Change-Id: Ibe82de48a837c38fadf812290776424ff7e9eddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614464
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158249}