Tabs directly check for widget activeness instead of plumbing through TabStrip.
It makes no sense for tabs (or the NTB, or, especially, the BrowserRootView) to ask the tabstrip if they should paint as if the widget is active - that is something between them and the widget, and the tabstrip doesn't enter into it.

This is a smol first step towards fixing tab colors in GM3.

Change-Id: I56e7bb3b186059eb9382f15e14c88aaece440bed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4616946
Commit-Queue: Taylor Bergquist <tbergquist@chromium.org>
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1158265}