Remove incorrect and obsolete VoiceRecognitionHandler test case.
I came across this test case when i was working on migrating VRH
Integration tests to Unit test suite.

The test was formerly passing for wrong reasons, as it was suppressed
by permission setting. The test was never verifying whether any tabs
are available - and it never needed to, as Tabs are not critical for
Voice Recognition Handler to be available.

Change-Id: I74dafc1f15956bd1409017b1313e98ceb03361d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4617055
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Auto-Submit: Tomasz Wiszkowski <ender@google.com>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1158277}