webauthn: require up-to-date JSON serialisation on Android.
This this change, passkey providers on Android can no longer send `null`
for optional values, and must return the easy-attestation values. I've
tweaked the code a little so that a public key need only be provided for
P-256 and Ed25519. Previously a public key had to be provided for all
algorithms that Chromium supported, but that seems a bit sharp if we add
more algorithms in the future.

Remote desktop uses are not yet ready for these policies to be applied,
so this change doesn't touch the CRD case at all (although we can likely
get away with the easy-accessor stuff). But since we need to
differentiate Android vs CRD cases in the code anyway, touching CRD
would add risk to a change that we mainly want to get established on
Android.

Bug: 1427843
Change-Id: I6488b952fe751f6e42f3bb12c53e92f5e1268934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4609977
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Auto-Submit: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1158286}