[Nearby Presence] Persist registration completed boolean
The LocalDeviceDataProvider now considers `IsRegistrationComplete` to be
true when the user name and profile url are persisted to prefs, as well
as checking a newly added pref boolean "registration_complete" to be true.
The new pref boolean is persisted when the entire first time registration
flow has been completed. This is a change from only when the server part
is completed. This will prevent "half-registered" states where consumers
think registraton is complete, when only the server portion is complete.
This will be used in a follow up CL that adds a CredentialManagerFactory
and will check that the factory is correctly registered before
executing the async callback.

Test: Unit tests only
Bug: b/276307539
Change-Id: I9dd4cf9228e8f7f443c6502d12643288ade1f39f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4558184
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#1158297}