[ios] Implement HTTPS-Upgrades
This CL implements the initial version of HTTPS-Upgrades on iOS. This
feature is essentially HTTPS-First Mode (formerly known as HTTPS-Only
Mode) without the HTTP interstitial, so it uses the same upgrade and
fallback logic as HTTPS-First Mode on iOS.

A brief summary of the HTTPS-Upgrades logic is as follows:

1. A class called HttpsOnlyModeUpgradeTabHelper which is both a
WebStatePolicyDecideder and WebStateObserver keeps track of navigations.

2. When a navigation starts, an instance of this class checks whether
it should allow the current navigation response. If the scheme of the
URL is https, or if the user allowed the HTTP navigation before by
clicking through the interstitial, or if the navigation isn't eligible
for an upgrade (e.g. it's localhost, an IP address etc), the navigation
is allowed and no upgrade is done.

3. Otherwise, the current response is canceled. A new navigation with
an HTTPS scheme is initiated and a timer is started.

4. The tab helper checks the upgraded navigation's response. If the
response is for a good HTTPS URL, the navigation is allowed and the
upgrade is successful.

5. Otherwise, the response is a broken or slow HTTPS or an HTTP URL.
The upgraded navigation is canceled. A fallback navigation is
started to the fallback HTTP URL. (In HTTPS-First Mode, this step
would show the HTTP interstitial instead).

Bug: 1449050
Change-Id: Ic45f7f1cf6862f0e090deecb6a211688dc8e4d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4575901
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158300}