Avoid passing NULL to memcmp
Some C++ libs legitimately use (nullptr, 0) as the empty slice for
the zeros vector. This can then end up passing NULL in the case of
memcmp(NULL, whatever, 0) which C frowns upon and may be annotated
as not allowing the NULL.

We can just ensure this doesn't happen by allocating an extra byte
in the vector.

Bug: 1322914
Change-Id: I9c38a6f0c91e54ba34e3a796a0ed6ffac2b293fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4607067
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Bob Beck <bbe@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158308}