[Tailored Security] Always add a TabModel observer on init of retry
The tab observer based retry mechanism for Tailored Security previously
worked by adding a TabModelList observer and waiting for a TabModel to
be added. In practice, there is frequently already a TabModel added,
which means that this implementation was adding the TabModelList
observer but often never successfully adding a TabModel observer because
another TabModel would not always be added.

This change addresses the problem by changing the retry initialization
logic to try adding a TabModel observer on the initial pass.

We also improve removing of TabModel observers by always adding the
ChromeTailoredSecurityService as a TabModelList observer. This is an
improvement because when a TabModel is removed, the TabModelList will
iterate over all of the TabModelList observers and call
OnTabModelRemoved().

* Add a RegisterObserver method which tries to add a TabModel
  observer before adding a TabModelListObserver.
* Update the ChromeTailoredSecurityService destructor to remove itself
  from the TabModelList observers after removing itself from the
  observers of the TabModel that it is observing.
* Change OnTabModelAdded() so that it:
    * does nothing if the ChromeTailoredSecurityService is already
      observing a TabModel
    * no longer removes itself from being a TabModelList observer (since
      if it is not a TabModelList observer, then it will miss running
      OnTabModelRemoved for any tabs that are removed.
* Start removing ChromeTailoredSecurityService from the TabModel
  observers of `observed_tab_model_`.
* Add RemoveTabModelListObserver and RemoveTabModelObserver methods so
  that the remove logic is encapsulated in a single method.
* Update the ChromeTailoredSecurityService's unit test suite to
  initialize some raw_ptr member variables to nullptr so that they are
  explicitly not initialized.

Change-Id: I642437b6178dd75a1332a0d74c3c2b6f9a9e53cd
Bug: 1452180,1448897
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4609457
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Javier Castro <jacastro@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158338}