Reland "[PolicyLogs] Adding chrome://policy/logs on Desktop"
chrome://policy/logs is already available on Android and enabled by default. This CL brings chrome://policy/logs to desktop behind a feature flag that is enabled by default.

The CL also adds a reference to the sequence that created the logger to  always post `AddLog()` to it if a different sequence adds a log. Some code paths previously tried to add a log from a different sequence, causing the DCHECKs to fail. This solution fixes the issue.

A POLICY_EXPORT is added for the LogHelper class to avoid dll linkage errors on Windows.

The ASAN Crash that caused the revert in crrev.com/c/4559846
was due to the last_dm_status_ being logged in "cloud_policy_client.cc" which has already been freed from the `NotifyClientError()` call.

Bug: b/282733597
Change-Id: Icb75e4cbb069eefea86958b65b3f957db295cdf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4567744
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1158355}