[Journeys] Pass takeover_action to Java if it's the only action present
The Android Journeys row relies on an action to trigger, and the
takeover_action was not previously passed to Java. For now it's simple
and effective to pass the takeover_action as a regular action; in the
future we may want to make this its own field.

Bug: 1455172
Change-Id: I01b9c55d67351752ea0b530bae9f9ae4e5836b26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4618728
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158361}