[Auto-disable Accessibility] Add AccessibilityManager instance to AccessibilityState
This CL continues the work to consolidate all accessibility logic in
Clank to a single source of truth.

With this CL we update the AccessibilityState to include a static member
variable for the AccessibilityManager. With each update to the services
we update the instance in case the Context has changed. The manager is
needed for clients to send AccessibilityEvents without needed to keep a
hard dependency on the AccessibilityManager, or having a reference to a
View.


AX-Relnotes: N/A
Low-Coverage-Reason: Added a pass-through method to a system service. This would not realistically catch bugs.
Bug: 1430202, b/265493191
Change-Id: Ib1ba25b545e3ea1688005b761c2fcd8357fccc30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4614407
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1158416}