DeskButton: Make button responsive to shelf overflow
This CL makes the desk button responsive to shelf overflow. Previously,
when the shelf was horizontal we would just force it to stay in expanded
state no matter what. This CL changes this behavior by adding a
condition where if the shelf will overflow in horizontal alignment with
the button in the expanded state, then we shrink the button, and vice
versa. It also adds and updates some tests to account for the changes.

Bug: b/272383056
Change-Id: I4ef107aa5b702a542ae7ec7818e945f568c8b92d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4606119
Reviewed-by: Yongshun Liu <yongshun@chromium.org>
Reviewed-by: Daniel Andersson <dandersson@chromium.org>
Commit-Queue: Ben Becker <benbecker@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158439}