Add content client API for legacy tech report.
When the legacy reporting is enabled by enterprise policy, pass the flag
into the renderer process to allow reporting being triggered.

Once the report is triggered, the report data will be passed back to the
browser process so that it can be uploaded to the enterprise management
service.


Bug: 1441215
Change-Id: Icb3935215a26fd34ff5e0fef28f7cf26fffa50d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4607147
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: Rick Byers <rbyers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158460}