snap-group: Let `SnapGroupController` decide to show or skip overview
This CL enables `SnapGroupController` to show or skip showing
overview on window snapped.
When showing overview is allowed, user is able to select a window
from the other side of the screen to complete a desired window
layout.
Whereas when showing overview is not allowed, overview will not
be allowed to show on window snapped. This will be useful when
restoring the window state for snap group.

Fixed: b/287092923
Test: Added unit test
Change-Id: I3718ca03928393c43b6d813032a36bdc808aca52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4606132
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158469}