Allow hiding popups for currently hidden notifications
This is necessary to prevent the user from being bombarded with popups
after the Welcome Tour.

Currently, all callers of `MarkSinglePopupAsShown()` either check for
visibility or can only be called on visible notifications, e.g., the
drag action for notifications. Therefore this is not a logical change
for current callers.

Also, this seems to have been a restriction that accidentally crept in
from a refactor nearly a decade ago, not a purposeful limitation.

Bug: b:277091643
Change-Id: I3e1ba50ffffc81aee7efbc61eb7abb604c094df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4616020
Quick-Run: Angus McLean <angusmclean@google.com>
Reviewed-by: Andre Le <leandre@chromium.org>
Commit-Queue: Angus McLean <angusmclean@google.com>
Cr-Commit-Position: refs/heads/main@{#1158474}