[Fixit] DownloadForegroundServiceManagerTest
DownloadForegroundServiceManagerTest was failing on multiple
builders. I reproduced on AVD generic_android31.textpb config and found
that if (VERSION.SDK_INT < VERSION_CODES.S) started evaluating to false
instead of true. The tests seem to assume that canStartForeground
will always return true and don't test that logic specifically
(which I would expect to be a unit test anyway). I've added in
an override to MockDownloadForegroundServiceManager to always
return true for the purposes of these tests. The tests are passing
on generic_android31.textpb AVD config and I expect this will fix the
issue on the emulators the test is running on.

Bug: 1297682
Change-Id: I883036718e82c3784374dbce464bfad79692bbf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4617059
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158490}