[VC] getting permission from every render frame host.
Some tabs can have permissions at their iframe, but not itself; in
case the tab still be able to access camera and microphone, but we
will not show the vc tray.

The fix is to go through each render frame host and aggregate the
permission of them all.


BUG=b:287150755

TEST="autoninja -C  out/Test chrome/test:browser_tests && out/Test/browser_tests --gtest_filter=*VideoConference*"

Change-Id: I5c81d1fd520bb58dee0b9a6c95ffb10596eaeff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4612926
Commit-Queue: Charles Zhao <charleszhao@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158527}