[Nearby Share] Revert breaking changes to NearbyProcess creation.
Effectively revert crrev.com/c/4496546 (key word "effectively"). This CL
reverts the changes that crrev.com/c/4496546 made to NearbyProcessManagerFactory
and NearbyDependenciesProviderFactory.

crrev.com/c/4496546 changed how the Nearby utility process and the Nearby Share
+ Phone Hub features are instantiated. Before the breaking CL, just 1 utility
process and 1 keyed service per feature was instantiated. After the breaking Cl,
more than 1 of each is created -- causing bugs like b/287061553.

We are unable to actually revert crrev.com/c/4496546 because doing so causes
regression in OOBE, and we're unable to understand why. Because this fix needs
to be land ASAP and be merged back to 115, we chose to revert just the files
that we know to be causing b/287061553.

b/287061553 is a breaking regression in the Nearby Share feature.

Test:
  - Reproduced b/287061553 before applying this CL.
  - Did not reproduce b/287061553 after applying this CL.
  - All unit tests passed; adjusted Quick Start tests to be behind the QS feature flag.

Bug: b/287061553
Change-Id: If5e11a3c8870b21703fef49b2175e32a22686db6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4618867
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Joe Antonetti <joeantonetti@google.com>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158538}