Make crosapi web app browser tests use LacrosOnly
This is in preparation for removing the kWebAppsCrosapi flag as well as
for removing the LacrosSideBySide and LacrosPrimary modes.

Since ChromeOS in LacrosOnly mode does not open an Ash browser window
but tests still rely on browser(), make tests create an Ash browser
window. This is meant to be a temporary workaround.

It turns out that a few tests that were running with kWebAppsCrosapi
are actually failing with LacrosOnly enabled for deeper reasons. These
tests are being disabled in this CL and bugs were filed to kick off
further investigation.

Bug: chromium:1440448, chromium:1450158
Bug: b:287165243, b:287166490, b:287165206
Bug: b:287166176, b:287166775, b:287167145
Change-Id: Ie113994a1dd8cf93921740c4f9dc8ba3b5ffd551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4592930
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158594}