[WinCameraZeroCopy] Manage the lifetime of IMFBuffer
This CL is a follow-up patch of CL:4323337. It aims to manage the
lifetime of IMFBuffer.

1) Extend CapturedExternalVideoBuffer to hold a ComPtr<IMFBuffer>.

2) Change VideoCaptureBufferPool::ReserveIdForExternalBuffer to take
in a CapturedExternalVideoBuffer instead of just a
GpuMemoryBufferHandle (the external buffer also contains the GMB handle
inside it).

3) Add a Windows-only
VideoCaptureBufferTracker::SetIMFBuffer(ComPtr<IMFBuffer>) method
that's called by ReserveIdForExternalBuffer when calling
SetHeldByProducer(true) i.e. when creating a new tracker or reusing an
existing tracker. The implementation in the Windows
GpuMemoryBufferTracker will just store the ComPtr as a field - this
ensures the IMFBuffer cannot be recycled by MediaFoundation.

4) Override OnHeldByConsumersChanged() in the Windows
GpuMemoryBufferTracker so that it drops the ComPtr<IMFBuffer> when held
by consumers becomes false - that way the IMFBuffer is free to be
recycled by MediaFoundation.

5) CL keeps storing member variables `GpuMemoryBufferHandle` for
GpuMemoryBufferTracker(win) since GpuMemoryBufferTrackerMac must need
store member variable `handle.io_surface`, so the common
interface(CreateTracker) needs to pass gmb handle and keep storing
variable.

Bug: 1120900
Change-Id: Icf809978786b3d3c5b18ca9fd1adb0a161d561f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4571114
Commit-Queue: Shuhai Peng <shuhai.peng@intel.com>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158618}