[FilesBulkPinning] Extract Speedometer to common place
The Speedometer is currently being used by the IOTask framework to work
out remaining time for a file transfer. For bulk pinning this
calculation should be reused to identify total time to pin all the
files, however, chromeos/components can't depend on chrome/* so move the
speedometer files to a location that can be used.

Bug: b:279847004
Test: chromeos_unittests --gtest_filter=*Speedometer*
Change-Id: I4303b8d65c1ca6fa477530125c789919f2c80f66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4617787
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1158639}